requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiokafka==0.7.1
    # via -r requirements.in
aiopg==1.2.1
    # via -r requirements.in
aiosqlite==0.17.0
    # via -r requirements.in
amqp==5.0.6
    # via kombu
async-timeout==3.0.1
    # via aiopg
billiard==3.6.4.0
    # via celery
celery==5.1.1
    # via -r requirements.in
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
greenlet==1.1.0
    # via sqlalchemy
kafka-python==2.0.2
    # via
    #   -r requirements.in
    #   aiokafka
kombu==5.1.0
    # via celery
prompt-toolkit==3.0.19
    # via click-repl
psycopg2-binary==2.9.1
    # via aiopg
pytz==2021.1
    # via celery
six==1.16.0
    # via click-repl
sqlalchemy==1.4.18
    # via -r requirements.in
typing-extensions==3.10.0.0
    # via aiosqlite
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
